# user-order-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: User Order API
  version: 1.0.0
  description: 使用 MySQL 持久化、短期 JWT 与可撤销刷新会话的用户订单 API。
servers:
  - url: http://localhost:8888/api/v1
    description: 本地开发环境
security:
  - BearerAuth: []
paths:
  /health:
    get:
      security: []
      summary: 健康检查
      responses:
        '200':
          description: 服务可用
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/register:
    post:
      security: []
      summary: 注册普通用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          $ref: '#/components/responses/AuthResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/EmailAlreadyExists'
  /auth/login:
    post:
      security: []
      summary: 邮箱密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthResult'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
  /auth/refresh:
    post:
      security: []
      summary: 使用 HttpOnly Refresh Cookie 轮换会话
      responses:
        '200':
          $ref: '#/components/responses/AuthResult'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /auth/logout:
    post:
      security: []
      summary: 撤销当前刷新会话并清除 Cookie
      responses:
        '204':
          description: 已退出登录
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /auth/me:
    get:
      summary: 当前登录用户
      responses:
        '200':
          description: 当前用户公开资料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /users:
    get:
      summary: 查询用户列表（仅 admin）
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/AfterID'
      responses:
        '200':
          $ref: '#/components/responses/UserList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: 创建无登录凭据的用户（仅 admin；普通注册使用 /auth/register）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{id}:
    get:
      summary: 按 ID 查询用户（本人或 admin）
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: 用户详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UserNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders:
    get:
      summary: 查询订单列表（普通用户仅看到自己的订单）
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/AfterID'
      responses:
        '200':
          $ref: '#/components/responses/OrderList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: 创建订单（普通用户自动绑定当前用户）
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: 可选的创建幂等键；去除首尾空白后不能为空，最长 255 个字符。同键同 userId/amount 返回首次创建的订单。
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: 幂等重放成功，返回首次创建的订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '409':
          $ref: '#/components/responses/IdempotencyKeyConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{id}:
    get:
      summary: 按 ID 查询订单
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{id}/pay:
    post:
      summary: 确认订单支付
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: 支付成功，或订单已是 paid 状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/InvalidOrderState'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{id}/cancel:
    post:
      summary: 取消订单
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: 取消成功，或订单已是 cancelled 状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '409':
          $ref: '#/components/responses/InvalidOrderState'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    AfterID:
      name: afterId
      in: query
      schema:
        type: integer
        format: int64
        minimum: 1
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: Ada
        email:
          type: string
          format: email
          example: ada@example.com
        password:
          type: string
          format: password
          minLength: 12
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
    AuthResult:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
          description: 仅保存在前端内存，并通过 Authorization Bearer 发送；不要写入 localStorage。
        user:
          $ref: '#/components/schemas/AuthUser'
    CreateUserRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: Ada
        email:
          type: string
          format: email
          example: ada@example.com
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Ada
        email:
          type: string
          format: email
          example: ada@example.com
        createdAt:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required: [amount]
      properties:
        userId:
          type: integer
          format: int64
          minimum: 1
          description: admin 可为其他用户指定；普通用户可省略，服务端自动使用当前用户。
          example: 1
        amount:
          type: integer
          format: int64
          minimum: 1
          description: 最小货币单位。
          example: 2599
    Order:
      type: object
      required: [id, userId, amount, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 1
        amount:
          type: integer
          format: int64
          example: 2599
        status:
          type: string
          enum: [pending, paid, cancelled]
          example: pending
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: 面向调用方的稳定错误码，不应依赖 error 文案判断分支。
          example: USER_NOT_FOUND
        error:
          type: string
  responses:
    AuthResult:
      description: 返回短期 Access Token，并设置 HttpOnly、SameSite=Strict 的 refresh_token Cookie。
      headers:
        Set-Cookie:
          schema:
            type: string
          description: Refresh Cookie；浏览器自动保存，前端 JavaScript 不可读取。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthResult'
    Unauthenticated:
      description: 未带、无效、过期或已撤销的认证凭据，错误码为 UNAUTHENTICATED
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidCredentials:
      description: 登录凭据无效，错误码为 INVALID_CREDENTIALS
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    EmailAlreadyExists:
      description: 邮箱已注册，错误码为 EMAIL_ALREADY_EXISTS
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 已登录但无资源权限，错误码为 FORBIDDEN
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UserList:
      description: 用户游标分页结果
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items:
                type: array
                items:
                  $ref: '#/components/schemas/User'
              nextCursor:
                type: string
                description: 存在下一页时返回；作为下次请求的 afterId。
    OrderList:
      description: 订单游标分页结果
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
              nextCursor:
                type: string
                description: 存在下一页时返回；作为下次请求的 afterId。
    BadRequest:
      description: 参数或请求体不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UserNotFound:
      description: 用户不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    OrderNotFound:
      description: 订单不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    IdempotencyKeyConflict:
      description: 幂等键已被不同的创建请求使用，错误码为 IDEMPOTENCY_KEY_CONFLICT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidOrderState:
      description: 订单当前状态不允许该操作，错误码为 INVALID_ORDER_STATE
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnsupportedMediaType:
      description: Content-Type 不是 application/json
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: 未预期的服务端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'